fix(org): a major refactor
combined `<Anime />` and `<Manga />` components into one component
