--- src/content/anime/sass/Content.scss
@@ -1,13 +1,0 @@
-@import '../../../sass/index.scss';
-
-ul {
-    font-size: 14px;
-}
-
-.video-listing {
-
-    &:hover {
-        color: $blue5;
-    }
-
-}
