+++ src/content/sass/style.scss
@@ -31,3 +31,11 @@
     }
 
 }
+
+.video-listing {
+
+    &:hover {
+        color: $blue5;
+    }
+
+}
